requirements.txt
# Core dependencies
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Utilities
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyyaml>=6.0

# Development dependencies
pytest>=8.0.0
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0 